pyproject.toml
[project]
name = "cross-factory-tta"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "ultralytics>=8.3",
    "torch>=2.3,<2.11",
    "torchvision>=0.18",
    "opencv-python>=4.9",
    "pillow>=10.0",
    "pyyaml>=6.0",
    "tqdm>=4.66",
    "numpy>=1.26",
    "kagglehub>=0.3",
    "transformers>=5.5.4",
    "accelerate>=1.13.0",
    "timm>=1.0.26",
    "safetensors>=0.7.0",
    "pycocotools>=2.0.11",
]

[dependency-groups]
dev = []

[[tool.uv.index]]
name = "pytorch-cu128"
url = "https://download.pytorch.org/whl/cu128"
explicit = true

[tool.uv]
concurrent-downloads = 1
cache-dir = "/mnt/code/.uv-cache"

[tool.uv.sources]
torch = { index = "pytorch-cu128" }
torchvision = { index = "pytorch-cu128" }

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
